Внес изменения, некоторые пропустил, так как повлияли на работоспособность
