+++ src/main/java/ru/yandex/practicum/filmorate/dao/FilmStorage.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.dao;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import javax.xml.crypto.Data;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -25,5 +26,5 @@ public interface FilmStorage {
     Map<Long, Film> getFilms();
 
     //возвращает список фильмов с режиссером и сортировкой по году и количеству лайков
-    List<Film> getListFilmsByDirector(int id, Map<Data, String> sortBy);
+    Collection<Film> getFilmsByDirector(int directorId, Collection<String> sortBy);
 }
